+++ packages/core/src/components/legend-item/legend-item.interfaces.tsx
@@ -1,3 +1,13 @@
-export interface LegendItemRemoveClickEvent {
-  originalEvent: MouseEvent;
+export interface LegendItemActiveChangeEvent {
+  /**
+   * De huidige status van de Legend Item.
+   */
+  current: boolean;
+
+  /**
+   * De gewenste status van de Legend Item.
+   */
+  next: boolean;
+
+  originalEvent: Event;
 }