+++ packages/dso-toolkit/src/components/map-controls/map-controls.demo.ts
@@ -1,5 +1,6 @@
 export const mapControlsDemoCss = `
   .symboolcode[data-symboolcode=regelingsgebied] {
     background-image: url(images/regelingsgebied.png);
+    background-size: contain;
   }
 `;